Store interned strings only once
- the `Vec` for id -> string lookups now stores `*const str`s
- `Symbol` now implements `Deref<Target=str>`, simplifying usage

Additionally, the `HashMap` for string -> id lookups is now a `HashSet`.
This means lookups retain a reference to the key, eventually enabling
`Symbol` to store a direct pointer to the string and eliminating the `ids`
field entirely.

This work is left undone because I haven't been able to create the keyword
symbol consts as raw pointers to length-prefixed strings.