Track local variable initialization dynamically
The initial SSA construction did not handle uninitialized variables
correctly, whether being read or written to, when their uninitialization
was dynamic.

This adds a separate flag for each local to track whether it has been
intialized, and removes any logic around uses of `Undef`. In the future
those optimizations will instead be done as part of normal constant
propagation and dead code removal.