Implement `with` loops
This separates "scopes" from "entities." An entity is a single internal id
with a set of fields; a scope is a GML-facing id that *may* refer to
multiple entities (e.g. self, other, all, object ids).

Reads from scope fields give the value from the first entity; writes to
scope fields write to all entities.

Some cases are known to use only single-entity scopes (self, other, global,
and unprefixed member/globalvar access). These are currently recognized in
the frontend.

This also introduces the `Project` SSA instruction, to allow the `With`
instruction to provide two values. Register allocation for `Project` is
currently a hack, assuming instructions are never reordered and that a
`With`'s only `Project` instructions are placed immediately following it.