Make `vm::Value` non-`Copy`
This moves us closer to a sound and safe API as discussed in #40. Because
a `vm::Value` may point to an array, users must check its type and perhaps
increment its refcount anywhere they duplicate the value.

As part of overhauling `vm::Value`, this commit also slightly tweaks the
NaN-boxing scheme. The quiet bit is now always set, moving tagged values
up to the top of the available bit-pattern range. This simplifies decoding
such that all real values are contiguous. (See #43)

To avoid introducing new, redundant refcount traffic, there are now also
`vm::ValueRef` and `vm::ArrayRef` types that share a representation with
their owned counterparts, but can only exist as a borrow.

The interpreter relies entirely on codegen to provide it with sound
bytecode that does not violate host language rules. This means that each
opcode handler must carefully choose whether to borrow, move, or clone its
operands, so that codegen can rely on that behavior.