Replace `ErrorHandler` trait with direct use of `ErrorPrinter`
The original reason for the trait, redirecting error messages to various
modes of output, is now accomplished using a different mechanism. The
concrete `ErrorPrinter` type now implements more functionality common to
all output modes.

This commit also lifts `ErrorPrinter` out of the `front` module as it is
now also used by the `vm` module and the `engine` crate, and replaces some
uses of `format!` with `format_args!` to reduce allocations.