Speed up type checking for generated bindings
Incremental builds of the runner crate spend most of their time type
checking the generated `register` functions, and this will only grow as
more APIs are filled out.

The slowest part of this process appears to be inferring the arguments to
each `vm::Bind` and finding its associated impl of `vm::FnBind`. Running a
clean `cargo check` for the `runner` crate previously took 7.81s. Doing
this type inference only once per API brings this to 4.21s; specifying the
type arguments we do know brings it to 3.04s.